ircd: Move IRCd sources to ircd subfolder

+++ ircd/client_send.cpp
@@ -16,12 +16,12 @@ void Client::send_motd()
   const std::string& motd = server.get_motd();
   size_t prev = 0;
   size_t next = motd.find("\n");
-  
+
   while (next != motd.npos)
   {
     len = snprintf(buffer + total, sizeof(buffer) - total,
         ":%s 372 %s :%.*s\r\n",
-        server.name().c_str(), nick().c_str(), next-prev, &motd[prev]);
+        server.name().c_str(), nick().c_str(), (int) (next-prev), &motd[prev]);
     total += len;
     prev = next + 1;
     next = motd.find("\n", prev);
@@ -56,7 +56,7 @@ void Client::send_lusers()
   send(RPL_LUSEROP,       std::to_string(server.get_counter(STAT_OPERATORS)) + " :operator(s) online");
   send(RPL_LUSERCHANNELS, std::to_string(server.get_counter(STAT_CHANNELS)) + " :channels formed");
   send(RPL_LUSERME, ":I have " + std::to_string(server.get_counter(STAT_LOCAL_USERS)) + " clients and 1 servers");
-  
+
   std::string mu = std::to_string(server.get_counter(STAT_MAX_USERS));
   std::string tc = std::to_string(server.get_counter(STAT_TOTAL_CONNS));
   send(250, "Highest connection count: " + mu + " (" + mu + " clients) (" + tc + " connections received)");
@@ -67,7 +67,7 @@ void Client::send_modes()
   char data[128];
   int len = snprintf(data, sizeof(data),
     ":%s MODE %s +%s\r\n", nickuserhost().c_str(), nick().c_str(), mode_string().c_str());
-  
+
   send_raw(data, len);
 }
 
@@ -75,11 +75,11 @@ void Client::send_stats(const std::string& stat)
 {
   char buffer[128];
   int  len;
-  
+
   if (stat == "u")
   {
     static const int DAY = 3600 * 24;
-    
+
     auto uptime = server.uptime();
     int days = uptime / DAY;
     uptime -= days * DAY;
@@ -87,7 +87,7 @@ void Client::send_stats(const std::string& stat)
     uptime -= hours * 3600;
     int mins = uptime / 60;
     int secs = uptime % 60;
-    
+
     len = snprintf(buffer, sizeof(buffer),
           ":%s %03u %s :Server has been up %d days %d hours, %d minutes and %d seconds\r\n",
           server.name().c_str(), RPL_STATSUPTIME, nick().c_str(), days, hours, mins, secs);
